pipelineの実行部分を複数ファイルに切り出し
