+++ Makefile
@@ -28,6 +28,8 @@ SRCS := main.c \
 		exec_redirect_set.c \
 		exec_redirect_out.c \
 		exec_pipeline.c \
+		exec_pipeline_fork.c \
+		exec_pipeline_wait.c \
 		exec_syscall.c \
 		err_main.c \
 		err_parser.c \
